* Fixed unit test   * Making chroot pulling in a fresh environment   * Adding progress dialog support for target creation   * making target filesystem mount operations copy over the host systems     basic network resolution files   * Fixed bug where deleting a project with multiple targets would     throw an exception   * Enabled project deletion to use a progress dialog   * Adding 'chroot-project' and 'chroot-target' command line options   * Adding a call to mount the target filesystem when clicking on the     target terminal icon   * making all target filesystems bind mount the host /tmp directory   * fixing a bug in mcaslin makefile such that only *.fset files are     installed   * Adding a change to make the target filesystem bind mount the host     systems apt cache

+++ libs/SDK.py
@@ -267,10 +267,20 @@ class SDK(object):
         # first delete all contained targets
         proj = self.projects[project_name]
         for target in proj.targets:
-            proj.delete_target(target)
+            proj.delete_target(target, False)
+        proj.targets.clear()
         # and then deal with the project
         proj.umount()
-        shutil.rmtree(os.path.join(proj.path))
+        cmd = "rm -fR %s" % (os.path.join(proj.path))
+        proc = subprocess.Popen(cmd.split(), stderr = subprocess.PIPE)
+        while proc.poll() == None:
+            try: 
+                self.cb.iteration(process=proc)
+            except:
+                pass
+        if proc.returncode != 0:
+            print >> sys.stderr, "ERROR: Unable to delete %s!" % (project_name)
+            raise ValueError(" ".join(proc.stderr.readlines()))
         os.unlink(os.path.join(self.config_path, proj.name + '.proj'))
         
     def __str__(self):
